Moving lambdas to BindingBase & DataFactory to sofapython3 library
