+++ CMakeLists.txt
@@ -78,7 +78,8 @@ set(HEADER_FILES
     src/SofaPython3/DataCache.h
     src/SofaPython3/DataHelper.h
     src/SofaPython3/PythonDownCast.h
-)
+    src/SofaPython3/DataFactory.h
+    )
 
 set(SOURCE_FILES
     src/SofaPython3/initModule.cpp
@@ -88,7 +89,8 @@ set(SOURCE_FILES
     src/SofaPython3/DataCache.cpp
     src/SofaPython3/DataHelper.cpp
     src/SofaPython3/PythonDownCast.cpp
-    )
+    src/SofaPython3/DataFactory.cpp
+)
 
 set(EXTRA_FILES
     SofaPython3Config.cmake.in
