+++ bindings/Sofa/src/SofaPython3/Sofa/Core/Binding_Base.h
@@ -38,6 +38,12 @@ public:
     /// Set the data field value from the array.
     static void SetDataFromArray(BaseData* data, const py::array& value);
     static bool SetData(BaseData* data, pybind11::object value);
+
+    static void addData(py::object py_self, const std::string& name, py::object value = py::object(), const std::string& help = "", const std::string& group = "Property", std::string type = "");
+    static void addDataFromData(Base* self, py::object d);
+    static py::list __dir__(Base* self);
+    static py::object __getattr__(py::object self, const std::string& s);
+    static void __setattr__(py::object self, const std::string& s, py::object value);
 };
